requirements.txt
fastapi>=0.104.0
uvicorn>=0.23.2
python-jose>=3.3.0
passlib>=1.7.4
python-multipart>=0.0.6
pydantic>=2.4.2
pydantic-settings>=2.0.3
azure-devops>=7.1.0b4
azure-mgmt-security>=5.0.0
azure-mgmt-resource>=23.0.1
azure-identity>=1.14.1
pandas>=2.1.1
PyYAML>=6.0.1
pytest>=7.4.2
pytest-cov>=4.1.0
requests>=2.31.0
python-dotenv>=1.0.0
aiohttp>=3.8.5
sqlalchemy>=2.0.21
alembic>=1.12.0
python-dateutil>=2.8.2
tenacity>=8.2.3